[Cache] deprecate all PSR-16 adapters, provide Psr16Cache instead
