+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -18,7 +18,7 @@
     "require": {
         "php": "^7.1.3",
         "ext-xml": "*",
-        "symfony/cache": "~4.2",
+        "symfony/cache": "~4.3",
         "symfony/config": "~4.2",
         "symfony/contracts": "^1.0.2",
         "symfony/dependency-injection": "^4.2",
