+++ src/Symfony/Component/Cache/Exception/CacheException.php
@@ -14,6 +14,12 @@ namespace Symfony\Component\Cache\Exception;
 use Psr\Cache\CacheException as Psr6CacheInterface;
 use Psr\SimpleCache\CacheException as SimpleCacheInterface;
 
-class CacheException extends \Exception implements Psr6CacheInterface, SimpleCacheInterface
-{
+if (interface_exists(SimpleCacheInterface::class)) {
+    class CacheException extends \Exception implements Psr6CacheInterface, SimpleCacheInterface
+    {
+    }
+} else {
+    class CacheException extends \Exception implements Psr6CacheInterface
+    {
+    }
 }
