+++ src/Symfony/Component/Cache/Exception/InvalidArgumentException.php
@@ -14,6 +14,12 @@ namespace Symfony\Component\Cache\Exception;
 use Psr\Cache\InvalidArgumentException as Psr6CacheInterface;
 use Psr\SimpleCache\InvalidArgumentException as SimpleCacheInterface;
 
-class InvalidArgumentException extends \InvalidArgumentException implements Psr6CacheInterface, SimpleCacheInterface
-{
+if (interface_exists(SimpleCacheInterface::class)) {
+    class InvalidArgumentException extends \InvalidArgumentException implements Psr6CacheInterface, SimpleCacheInterface
+    {
+    }
+} else {
+    class InvalidArgumentException extends \InvalidArgumentException implements Psr6CacheInterface
+    {
+    }
 }
