+++ src/Symfony/Component/Cache/Simple/ApcuCache.php
@@ -13,6 +13,11 @@ namespace Symfony\Component\Cache\Simple;
 
 use Symfony\Component\Cache\Traits\ApcuTrait;
 
+@trigger_error(sprintf('The "%s" class is deprecated since Symfony 4.3, use "%s" and type-hint for "%s" instead.', ApcuCache::class, ApcuAdapter::class, CacheInterface::class), E_USER_DEPRECATED);
+
+/**
+ * @deprecated since Symfony 4.3, use ApcuAdapter and type-hint for CacheInterface instead.
+ */
 class ApcuCache extends AbstractCache
 {
     use ApcuTrait;
