+++ src/Symfony/Component/Cache/Tests/Simple/ApcuCacheTest.php
@@ -13,6 +13,9 @@ namespace Symfony\Component\Cache\Tests\Simple;
 
 use Symfony\Component\Cache\Simple\ApcuCache;
 
+/**
+ * @group legacy
+ */
 class ApcuCacheTest extends CacheTestCase
 {
     protected $skippedTests = [
