+++ src/Symfony/Component/Cache/Tests/Simple/ArrayCacheTest.php
@@ -15,6 +15,7 @@ use Symfony\Component\Cache\Simple\ArrayCache;
 
 /**
  * @group time-sensitive
+ * @group legacy
  */
 class ArrayCacheTest extends CacheTestCase
 {
