+++ src/Symfony/Component/Cache/Tests/Simple/ChainCacheTest.php
@@ -19,6 +19,7 @@ use Symfony\Component\Cache\Simple\FilesystemCache;
 
 /**
  * @group time-sensitive
+ * @group legacy
  */
 class ChainCacheTest extends CacheTestCase
 {
