+++ src/Symfony/Component/Cache/Tests/Simple/FilesystemCacheTest.php
@@ -16,6 +16,7 @@ use Symfony\Component\Cache\Simple\FilesystemCache;
 
 /**
  * @group time-sensitive
+ * @group legacy
  */
 class FilesystemCacheTest extends CacheTestCase
 {
