+++ src/Symfony/Component/Cache/Tests/Simple/PdoCacheTest.php
@@ -16,6 +16,7 @@ use Symfony\Component\Cache\Tests\Traits\PdoPruneableTrait;
 
 /**
  * @group time-sensitive
+ * @group legacy
  */
 class PdoCacheTest extends CacheTestCase
 {
