+++ src/Symfony/Component/Cache/Tests/Simple/Psr6CacheTest.php
@@ -13,6 +13,9 @@ namespace Symfony\Component\Cache\Tests\Simple;
 
 use Symfony\Component\Cache\Simple\Psr6Cache;
 
+/**
+ * @group legacy
+ */
 abstract class Psr6CacheTest extends CacheTestCase
 {
     protected $skippedTests = [
