+++ src/Symfony/Component/Cache/Tests/Simple/Psr6CacheWithoutAdapterTest.php
@@ -15,6 +15,7 @@ use Symfony\Component\Cache\Tests\Fixtures\ExternalAdapter;
 
 /**
  * @group time-sensitive
+ * @group legacy
  */
 class Psr6CacheWithoutAdapterTest extends Psr6CacheTest
 {
