+++ src/Symfony/Component/Cache/Tests/Simple/RedisArrayCacheTest.php
@@ -11,6 +11,9 @@
 
 namespace Symfony\Component\Cache\Tests\Simple;
 
+/**
+ * @group legacy
+ */
 class RedisArrayCacheTest extends AbstractRedisCacheTest
 {
     public static function setupBeforeClass()
