+++ src/Symfony/Component/Cache/Tests/Simple/RedisCacheTest.php
@@ -13,6 +13,9 @@ namespace Symfony\Component\Cache\Tests\Simple;
 
 use Symfony\Component\Cache\Simple\RedisCache;
 
+/**
+ * @group legacy
+ */
 class RedisCacheTest extends AbstractRedisCacheTest
 {
     public static function setupBeforeClass()
