+++ src/Symfony/Component/Cache/Tests/Simple/RedisClusterCacheTest.php
@@ -11,6 +11,9 @@
 
 namespace Symfony\Component\Cache\Tests\Simple;
 
+/**
+ * @group legacy
+ */
 class RedisClusterCacheTest extends AbstractRedisCacheTest
 {
     public static function setupBeforeClass()
