+++ src/Symfony/Component/Cache/Tests/Simple/TraceableCacheTest.php
@@ -16,6 +16,7 @@ use Symfony\Component\Cache\Simple\TraceableCache;
 
 /**
  * @group time-sensitive
+ * @group legacy
  */
 class TraceableCacheTest extends CacheTestCase
 {
