+++ src/Symfony/Component/Cache/composer.json
@@ -24,7 +24,6 @@
         "php": "^7.1.3",
         "psr/cache": "~1.0",
         "psr/log": "~1.0",
-        "psr/simple-cache": "^1.0",
         "symfony/contracts": "^1.0",
         "symfony/var-exporter": "^4.2"
     },
@@ -33,6 +32,7 @@
         "doctrine/cache": "~1.6",
         "doctrine/dbal": "~2.5",
         "predis/predis": "~1.1",
+        "psr/simple-cache": "^1.0",
         "symfony/config": "~4.2",
         "symfony/dependency-injection": "~3.4|~4.1",
         "symfony/var-dumper": "^4.1.1"
